Guess Textbox Gets Cleared After New Puzzle Loaded
The guess textbox previously did not get cleared when a new puzzle was loaded, so a word that was valid for the previous puzzle could still be in the textbox and would be invalid for the currently loaded word. Also moved some functions around and added comments to denote sections of code.