Fixed CLI Exit Requiring 2 Enters
When the `CLIController` would recive the exit command it would set the
exit flag to true, but then return to the begining of the loop and wait
for input to satisfy scanner.hasNextLine before checking the exit flag.
This has been fixed by removing the exit flag and simply breaking
the loop when the exit command is recived.